simple_agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "mcp>=1.0.0",
]
